fix(dynamodb): cannot use attribute as key in a GSI, non-key in another (#7075)
Removes the restriction that a key attribute cannot be used as a non-key
attribute in a distinct secondary index. This was excessively
restricting what can be done, as new integration tests demonstrated that
there is no problem in deploying such a configuration.

Also, switches from `Array`s to `Set`s to represent the sets of names that
must be unique.

Fixes #4398
