+++ packages/@aws-cdk/aws-dynamodb/package.json
@@ -63,15 +63,16 @@
   "license": "Apache-2.0",
   "devDependencies": {
     "@aws-cdk/assert": "0.0.0",
-    "@types/nodeunit": "^0.0.30",
+    "@types/jest": "^25.1.4",
     "aws-sdk": "^2.650.0",
     "aws-sdk-mock": "^5.1.0",
     "cdk-build-tools": "0.0.0",
     "cdk-integ-tools": "0.0.0",
     "cfn2ts": "0.0.0",
-    "nodeunit": "^0.11.3",
+    "jest": "^25.2.4",
     "pkglint": "0.0.0",
-    "sinon": "^9.0.1"
+    "sinon": "^9.0.1",
+    "ts-jest": "^25.3.0"
   },
   "dependencies": {
     "@aws-cdk/aws-applicationautoscaling": "0.0.0",
@@ -115,5 +116,6 @@
   },
   "awscdkio": {
     "announce": false
-  }
+  },
+  "jest": {}
 }
