+++ packages/@aws-cdk/aws-dynamodb/test/integ.dynamodb.mixed-key-gsi.expected.json
@@ -1,0 +1,83 @@
+{
+  "Resources": {
+    "TableCD117FA1": {
+      "Type": "AWS::DynamoDB::Table",
+      "Properties": {
+        "KeySchema": [
+          {
+            "AttributeName": "pkey",
+            "KeyType": "HASH"
+          }
+        ],
+        "AttributeDefinitions": [
+          {
+            "AttributeName": "pkey",
+            "AttributeType": "N"
+          },
+          {
+            "AttributeName": "foo",
+            "AttributeType": "S"
+          },
+          {
+            "AttributeName": "baz",
+            "AttributeType": "S"
+          },
+          {
+            "AttributeName": "bar",
+            "AttributeType": "S"
+          }
+        ],
+        "GlobalSecondaryIndexes": [
+          {
+            "IndexName": "IndexA",
+            "KeySchema": [
+              {
+                "AttributeName": "foo",
+                "KeyType": "HASH"
+              }
+            ],
+            "Projection": {
+              "NonKeyAttributes": [
+                "bar"
+              ],
+              "ProjectionType": "INCLUDE"
+            },
+            "ProvisionedThroughput": {
+              "ReadCapacityUnits": 5,
+              "WriteCapacityUnits": 5
+            }
+          },
+          {
+            "IndexName": "IndexB",
+            "KeySchema": [
+              {
+                "AttributeName": "baz",
+                "KeyType": "HASH"
+              },
+              {
+                "AttributeName": "bar",
+                "KeyType": "RANGE"
+              }
+            ],
+            "Projection": {
+              "NonKeyAttributes": [
+                "blah"
+              ],
+              "ProjectionType": "INCLUDE"
+            },
+            "ProvisionedThroughput": {
+              "ReadCapacityUnits": 5,
+              "WriteCapacityUnits": 5
+            }
+          }
+        ],
+        "ProvisionedThroughput": {
+          "ReadCapacityUnits": 5,
+          "WriteCapacityUnits": 5
+        }
+      },
+      "UpdateReplacePolicy": "Delete",
+      "DeletionPolicy": "Delete"
+    }
+  }
+}
